backend/requirements.txt
# API Framework
fastapi>=0.95.0,<0.96.0
uvicorn[standard]>=0.22.0,<0.23.0

# Database
tortoise-orm>=0.18.0,<0.19.0
asyncpg>=0.27.0,<0.28.0
aiosqlite>=0.17.0,<0.18.0

# Authentication
python-multipart>=0.0.9,<0.0.10
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.0.0,<4.2.0

# Data validation and settings
pydantic>=1.10.0,<2.0.0
python-dotenv>=0.21.0,<0.22.0

# HTTP client
httpx>=0.23.0,<0.24.0
requests>=2.28.0,<2.29.0

# Database migrations
aerich>=0.6.0,<0.7.0

# Utilities
iso8601>=0.1.16,<0.2.0
pypika-tortoise>=0.1.6,<0.2.0
tomlkit>=0.11.0,<0.12.0

# Logging
pytz>=2023.3